pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SamrahMukhtar", email = "samrah.mukhtar@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
hello-agent = "hello_agent:main"
my-agent = "hello_agent.agent_hello:my_first_agent"
Healthcare_agent = "hello_agent.moodmate:run_Healthcare"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
